+++ examples/c/03_inline_okl/main.c
@@ -3,6 +3,7 @@
 #include <math.h>
 
 #include <occa.h>
+#include <occa/c/experimental.h>
 
 //---[ Internal Tools ]-----------------
 // Note: These headers are not officially supported
