+++ examples/cpp/03_inline_okl/main.cpp
@@ -1,6 +1,7 @@
 #include <iostream>
 
 #include <occa.hpp>
+#include <occa/experimental.hpp>
 
 //---[ Internal Tools ]-----------------
 // Note: These headers are not officially supported
