+++ examples/cpp/08_arrays/main.cpp
@@ -1,6 +1,7 @@
 #include <iostream>
 
 #include <occa.hpp>
+#include <occa/experimental.hpp>
 
 //---[ Internal Tools ]-----------------
 // Note: These headers are not officially supported
