+++ include/occa.hpp
@@ -1,7 +1,6 @@
 #ifndef OCCA_HEADER
 #define OCCA_HEADER
 
-#include <occa/array.hpp>
 #include <occa/core.hpp>
 #include <occa/dtype.hpp>
 #include <occa/types.hpp>
