--- include/occa/array.hpp
@@ -1,6 +1,0 @@
-#ifndef OCCA_ARRAY_HEADER
-#define OCCA_ARRAY_HEADER
-
-#include <occa/array/array.hpp>
-
-#endif
