+++ include/occa/experimental.hpp
@@ -1,0 +1,9 @@
+#ifndef OCCA_EXPERIMENTAL_HEADER
+#define OCCA_EXPERIMENTAL_HEADER
+
+#include <occa/experimental/array.hpp>
+#include <occa/experimental/kernelBuilder.hpp>
+#include <occa/experimental/mpi.hpp>
+#include <occa/experimental/scope.hpp>
+
+#endif
