+++ include/occa/experimental/array.hpp
@@ -1,0 +1,6 @@
+#ifndef OCCA_EXPERIMENTAL_ARRAY_HEADER
+#define OCCA_EXPERIMENTAL_ARRAY_HEADER
+
+#include <occa/experimental/array/array.hpp>
+
+#endif
