+++ src/experimental/kernelBuilder.cpp
@@ -1,11 +1,11 @@
 #include <occa/core/device.hpp>
-#include <occa/core/kernelBuilder.hpp>
-#include <occa/core/scope.hpp>
 #include <occa/types/json.hpp>
 #include <occa/internal/core/kernel.hpp>
 #include <occa/internal/lang/kernelMetadata.hpp>
 #include <occa/internal/utils/lex.hpp>
 #include <occa/internal/utils/string.hpp>
+#include <occa/experimental/kernelBuilder.hpp>
+#include <occa/experimental/scope.hpp>
 
 namespace occa {
   //---[ kernelBuilder ]----------------
